libsnark: convert long long and unsigned long to C++11 fixed-width types

+++ src/snark/libsnark/algebra/curves/alt_bn128/alt_bn128_g1.cpp
@@ -10,8 +10,8 @@
 namespace libsnark {
 
 #ifdef PROFILE_OP_COUNTS
-long long alt_bn128_G1::add_cnt = 0;
-long long alt_bn128_G1::dbl_cnt = 0;
+int64_t alt_bn128_G1::add_cnt = 0;
+int64_t alt_bn128_G1::dbl_cnt = 0;
 #endif
 
 std::vector<size_t> alt_bn128_G1::wnaf_window_table;
